add test for combinePaths, fix mannwhitney_conf_int

+++ src/pyplier/combinePaths.py
@@ -2,4 +2,4 @@ import pandas as pd
 
 
 def combinePaths(*args):
-    return pd.concat(args, axis=1, join="outer").astype(float).fillna(0)
+    return pd.concat(args, axis=1, join="outer").fillna(0).astype(int)
